implemented output to buffer and _Z global
now there is a global opaque structure containing information about
the current zenroom environment saved inside lua space as _Z. This is
for now useful to save the stdout and stderr buffer (which this commit
implements) and in the future other zenroom specific pointers that
need to be processed globally by C callbacks. It also adds a new
global "CODE" containing the code being executed and exposes
init/teardown to host applications.