fix segfault on js build exit
this is an interesting case where a system free() call on a pointer
already freed by lua as lightuserdata on lua_close() is not leading to
a crash on any build but js builds (more sensitive to this).

caused indeed by saving the Z context inside L and freeing after
freeing L. Now should be OK, but good to keep an eye on teardown after
usage of deeped init/exec/teardown API takes place.