Started modularization of ECP at compile time
We cannot use the same runtime "factory" approach in ECDH here since
also types are changing (BIG numbers) so its not trivial in C. We are
confronted here with the same problem already faced by Milagro for
which code generation is the best solution. Further work will proceed
to modularize BIG and ECP* for more curves.