preliminary support for quadratic extension field
this is a vm representation of FP numbers which are basically BIG
number coordinates internal to ECPs whose excess is saved across
operations. This is a prototyping phase as eventually the type in
zenroom can be merged with bigs hiding a lot of complexity and the
existance of the big type.