removed signature verification on petition creation
this is a separate authentication channel from the petition and is
mostly in charge of the host application (authorization to create a
petition) no point in keeping, better lower complexity. It was
breaking determinism because of usage of JSON serialization which has
not deterministic sorting. Would be nice however to make JSON sorted.