introduced encapsulation header for base64
there is no deterministic way to distinguish a base64 from a string,
so a prefix is needed for this and in future more encapsulation
schemes. For this 4 chars are prefixed separated by colon from the
actual data (b64:). This introduces a regression for all previous data
created with Zenroom.

The feature comes after quite some piloting and realising that a
common mistake is to sum up multiple encapsulations especially using
the wrapper base64() inside map or standalone. This will put an end to
this ambiguity improving long-term DX.