fix: Add semihosting support for Zenroom to run on qemu
To use the semihosting will let us to test the Zenroom on
qemu-system-arm or such as cortex-m devices more easily.

But for running the semihosting correctly, we need to update the startup
file and the linker scripts.

This commit add a CMSIS repository into submodule for update the startup
file and the linker scripts to CMSIS and update the functions inside
cortex-m.c for semihosting support.

The Zenroom could be test on the following qemu scripts:
```
qemu-system-arm -M mps2-an385 -kernel src/zenroom.bin -semihosting -nographic
```