ci: Validate Arm Cortex-M targets
This commit adds the CI check for cortex-m targets.

The aps2-an386 needs the QEMU version > 5.0. But the defulat
QEMU version of the ubuntu's PPA is 4.2.0. So the target
changes from cortex-m4 to cortex-m3. When the ubuntu upgrade
the QEMU version, the cflags can change back from cortex-m3
to cortex-m4.

For testing purpose, the RAM size is increase as large as possible.