ci: Improve CI for cortex-m targets
This commit adds the zencode integration test and the crypto
integration test for Arm cortex-m targets.

For testing the Zenroom on the cortex-m3 device, the stderr and stdout
stream should be changed. The stderr messages in Zenroom will be
redirect to semihosting console. The stdout message will be written
into the "outlog" file since the QEMU doesn't have the stderr and stdout
stream.

Since the QEMU also can not receive the stdin stream, the test script
will be written into a temp file. The QEMU will open the temp file
via semihosting to get the test scripts.