Implement print stream for cortex-m target
This commit implements the output function for cortex-m target.
Since the QEMU didn't have the stderr and stdout stream. The
cortex-m device need to write the stderr message to semihosting
console via semihosting call. The messages of stdout will be written
into a "outlog" file.

The output will help CI to validate the cortex-m result in QEMU.