Really weird memory leak
After changing the body of the test with a loop that executed the test
multiple times, something really strange happened.

For a buffer of size 10 * 1024 the for loop executed until the end

For a buffer size of 512 * 1024 the number of iteration executed without
error is proportional to the size of the buffer.