feat: 🔃 foreach statement in zencode with tests
This commit introduces the foreach statement in zencode.
The following statements are supported:
- Foreach '' in ''
- Foreach '' in sequence from '' to '' with step ''

Foreach can be nested with if statements (that is there can be
a for inside a if and an if inside a for, if inside for inside if
are not supported since nested if do not exists).

Foreach can be nested, the result is a zip (elements are visited
in parallel).