fix: improve chain id import in ethereum transaction
Now chain id can also be imported from KEYS or DATA.
Chain ID can have encoding 'string' or 'integer'.

When the chain id is set, we look at ACK and see if the
variable references a value which is loaded. In this case:
- if the encoding is 'integer' his numeric value is imported.
- if the encoding is 'string' we try to import its numeric value,
  if it fails we import it as `O.from_string` (backward compatibility)

If the name is not loaded from keys or data, we use the value inside
the zencode script