feat: new conf scope=given to limit execution
when this conf setting is on only the given and then blocks are
processed, skipping all when as not known (as well if and foreach)
this allows caller applications to import data input before
processing.

work in progress: next will include the possibility to print out the
CODEC as a description of the imported input as well to mark input
that is not found without aborting the operation.