fix: be more permissive with unknown encodings
now Given accepts just the name of an object, be it a single element
or a dictionary, it will be decoded as-is without knowing much about
it, using the default codec.

moving this way it starts to make sense to change the default encoding
to string, which makes it possible to handle a lot more as default,
perhaps also implementing special conversion routine in When phase.