fix: keyring hidden but sizes shown, contents all zeroed
added OCTET.is_zero function for fast verification of contents, which
will also be printed as (all zero) in debugging for all values
matching this condition.

for BIG values in keyring (which may be force converted to OCTET in
future?) they are converted to a decimal value of their size, cheap
trick for now, but still makes it possible to know it in debug.